split up control and test
