+++ package.json
@@ -94,6 +94,7 @@
         "mermaid": "^9.1.5",
         "mobx": "^6.3.13",
         "node-fetch": "^2.6.1",
+        "posthog-node": "^2.0.2",
         "prism-react-renderer": "^1.2.0",
         "prismjs": "^1.25.0",
         "puppeteer-core": "^13.0.1",
